Refactored ProjectDetailsViewModels into separate classes
