+++ src/CompetitionPlatform/Models/ProjectViewModels/EditResultViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace CompetitionPlatform.Models.ProjectViewModels
+{
+    public class EditResultViewModel
+    {
+        public string ProjectId { get; set; }
+        public string UserId { get; set; }
+    }
+}
